refactor CYCCNT implementation as *example*
